src/senses/proximity/vl53l0x/distance_sensor.py
"""
<summary>
Controls the VL53L0X Time-of-Flight distance sensor for proximity detection
</summary>
<hardware>VL53L0X Time-of-Flight Distance Sensor (I2C)</hardware>
<dependencies>VL53L0X-python, smbus2</dependencies>
"""

from typing import Optional
import time
try:
    import VL53L0X
except ImportError:
    # Fallback for development environment
    VL53L0X = None


class DistanceSensor:
    """
    <summary>
    Manages VL53L0X distance sensor operations including ranging and calibration
    </summary>
    """
    
    def __init__(self, i2c_address: int = 0x29) -> None:
        """
        <summary>Initialize distance sensor with I2C configuration</summary>
        <param name="i2c_address">I2C address of the VL53L0X sensor</param>
        <returns>None</returns>
        """
        self.i2c_address = i2c_address
        self.sensor = None
        self.is_initialized = False
        self.measurement_mode = "single"
        self.timing_budget = 33000  # microseconds
    
    def initialize(self) -> bool:
        """
        <summary>Initialize VL53L0X sensor hardware and configuration</summary>
        <returns>True if successful, False otherwise</returns>
        """
        try:
            if VL53L0X:
                self.sensor = VL53L0X.VL53L0X(i2c_address=self.i2c_address)
                self.sensor.open()
                self.sensor.start_ranging(VL53L0X.Vl53l0xAccuracyMode.BETTER)
                self.is_initialized = True
                return True
            return False
        except Exception as e:
            print(f"Distance sensor initialization failed: {e}")
            return False
    
    def get_distance(self) -> Optional[float]:
        """
        <summary>Get distance measurement in millimeters</summary>
        <returns>Distance in mm or None if failed</returns>
        """
        if not self.is_initialized:
            return None
        
        try:
            # Skeleton implementation
            return 0.0
        except Exception as e:
            print(f"Distance measurement failed: {e}")
            return None
    
    def get_distance_cm(self) -> Optional[float]:
        """
        <summary>Get distance measurement in centimeters</summary>
        <returns>Distance in cm or None if failed</returns>
        """
        distance_mm = self.get_distance()
        if distance_mm is not None:
            return distance_mm / 10.0
        return None
    
    def is_object_detected(self, threshold_cm: float = 30.0) -> bool:
        """
        <summary>Check if object is detected within threshold distance</summary>
        <param name="threshold_cm">Detection threshold in centimeters</param>
        <returns>True if object detected, False otherwise</returns>
        """
        distance = self.get_distance_cm()
        if distance is not None:
            return distance <= threshold_cm
        return False
    
    def set_measurement_mode(self, mode: str) -> bool:
        """
        <summary>Set measurement mode (single or continuous)</summary>
        <param name="mode">Measurement mode ('single' or 'continuous')</param>
        <returns>True if successful, False otherwise</returns>
        """
        if mode in ["single", "continuous"]:
            self.measurement_mode = mode
            return True
        return False
    
    def set_timing_budget(self, budget_us: int) -> bool:
        """
        <summary>Set measurement timing budget in microseconds</summary>
        <param name="budget_us">Timing budget in microseconds</param>
        <returns>True if successful, False otherwise</returns>
        """
        if 20000 <= budget_us <= 1000000:  # Valid range
            self.timing_budget = budget_us
            return True
        return False
    
    def calibrate(self, reference_distance_mm: float) -> bool:
        """
        <summary>Calibrate sensor with known reference distance</summary>
        <param name="reference_distance_mm">Known reference distance in mm</param>
        <returns>True if successful, False otherwise</returns>
        """
        if not self.is_initialized:
            return False
        
        # Skeleton implementation
        return True
    
    def get_sensor_info(self) -> dict:
        """
        <summary>Get sensor information and status</summary>
        <returns>Dictionary containing sensor info</returns>
        """
        if not self.is_initialized:
            return {}
        
        # Skeleton implementation
        return {
            "address": self.i2c_address,
            "mode": self.measurement_mode,
            "timing_budget": self.timing_budget,
            "status": "ready"
        }
    
    def cleanup(self) -> None:
        """
        <summary>Clean up distance sensor resources</summary>
        <returns>None</returns>
        """
        if self.sensor:
            self.sensor.stop_ranging()
            self.sensor.close()
            self.sensor = None
        self.is_initialized = False