requirements.txt
# Zolo Robot Project Dependencies

# GPIO Control
RPi.GPIO==0.7.1
gpiozero==1.6.2

# NeoPixels
rpi_ws281x==4.3.4
adafruit-circuitpython-neopixel==6.3.4
adafruit-blinka==8.22.2

# Sensors
VL53L0X-python==0.1.3
smbus2==0.4.2
Adafruit-CircuitPython-TSL2561==3.3.4

# Camera & Vision
# Note: libcamera is system-level, install with: sudo apt install libcamera-tools
picamera2==0.3.12
opencv-python==4.8.1.78
numpy==1.24.3

# Audio Processing
pyaudio==0.2.11
sounddevice==0.4.6
speech-recognition==3.10.0
pyttsx3==2.90

# GUI (Touchscreen)
# Note: tkinter comes with Python standard library
kivy==2.2.0
# PyQt5==5.15.9  # Uncomment if needed

# Core Python Libraries
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
psutil==5.9.5

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development Tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1
isort==5.12.0

# Logging and Monitoring
colorlog==6.7.0

# Future AI Integration (Optional)
# bittensor==6.9.3  # Uncomment when ready for AI integration
# transformers==4.33.2  # Uncomment if using local AI models

# Additional CircuitPython Libraries
adafruit-circuitpython-register==1.9.16
adafruit-circuitpython-busdevice==5.2.6